Wait for spinner to disappear Attempt 2
This strategy is moot because on the SoundCloud river of songs, the
loading spinner never goes away until you reach the end of the page.
It is always there, and the web designers and devs of SC decided to
create an illusion that a loading spinner appears when songs are
loading and disappears when the songs have loaded. In fact, the loading
spinner is just the last element of the unordered <ul> list of songs,
and when the user scrolls to the bottom of the list to trigger the lazy
loading list to fetch more songs, they will see the spinner. When the
songs are fetched and appended to the list, the loading spinner is
pushed to the bottom of the page, off the screen, making it "disappear"

The loading spinner is only removed from the page when the end of the
page is reached and all the songs have been fetched. This script
already has logic to check for the eof element that signals the same
thing.

The only reason I want to commit this change is to keep a record of a
reasonable strategy for waiting for an element to appear and disappear,
since this would have worked in theory.