Check if lazy loading song list has stalled
As mentioned in the comment, the previous strategy was moot, so now the
number of songs in the list is the only metric to verify whether the
server has timeout, or for some other reason loading of the song list
has stalled. We don't want to "mindlessly" continue pressing PAGE_DOWN
or the END key if the lazy loading list has stopped loading and the
songs have stopped loading because that is not human-like behavior, and
it is a waste of resources.

There still isn't a heuristic for knowing if we are scrolling too
quickly, thus not allowing the lazy list to load the next batch of songs
before hitting one of the scroll down keys. So this is the best we can
do. However, it will be improved by adding more time to the pause
whenever we notice the lazy list has not loaded since the last time we
checked it. We should also be patient and not bail the first time this
happens, and try checking two or more times before pronouncing the feed as
"frozen." So a retry_counter could be used for that.

Finally, as the length of the list increases the amount of memory
consumed increases, so checking the length of the list will probably
have a high performance cost. Therefore one solution is to only check
it at certain intervals. Based on past execution of this script, it can
take up to 2000 scroll times to reach the bottom of the list. Memory
cost is noticeable after 1000 scrolls. And the cost to do the check
itself gets higher as you are asking Selenium to process the bulk of
the page each time. So we can define a list of checkpoints to stop at
when the scrollCount reaches them.