Only count songs at checkpoints to reduce overhead
As mentioned in the previous commit, the new strategy for checking if
the lazy loading list has stalled could have a high performance cost
when the song list gets very big. So with a predefined set of
checkpoints we will only do this test a handful of times instead of
thousands of times, with each check taking more resources. This is
only in theory, but from experience I have seen the scrolling of the
page take more and more time per scroll as loading of the songs takes
more time.

If a checkpoint test fails (number of songs has not changed since last
check) then the test will be performed again on the next loop cycle to
double check whether the loading has really stalled or if it was a
false alarm. The pause between scrolls/cycles increases by 1 second to
give the SoundCloud server more time to load. A max retries limits the
number of times we do a "double check." Finally, if the songs start to
load again, checkpoint_retries is reset so in case the next checkpoint
test fails, the script will not bail too soon; we want to perform the
test the same number of times before assuming the server has stalled.

The current values in the song_count_checkpoints set are small for
testing purposes, but in production they will be adjusted for a much
bigger scale. I decided to statically define the checkpoints instead
of mathematically making them, such as doubling or tripling the last
number to get the next checkpoint number. I want more control over it.