Revert "Remove first_art, reference no longer necessary"
This reverts commit e4ec50887980c2d354fa4cdf228bc06d57c4cd1b.
I found that the problem which I was trying to address originaly by
moving the cursor to the first song art and hovering near it, has
reoccurred: if the cursor is not over the main content of the page then
when the END key or PAGE_DOWN key is sent by Selenium to the browser it
has no effect. This makes sense because that is what happens when you
browse a web page yourself.

However, I'm not sure why this problem stopped occurring back when I
removed this code, or what made it no longer necessary. It works now
that I restored it, but the "firefox workaround" for the move_to_element
function seems necessary also for Chrome.