Expand checkpoint intervals, increase max retries
In the first test with the new "song count checkpoint" system, the song
count did not change after the 4th checkpoint (1000 scrollCount). It
failed again 4 more times consecutively and the script quit and saved
8010 songs. This is the designed behavior, but the checkpoint intervals
and max retry count should be adjusted so the conditions are not so
strict for assuming the server has stalled.

A second test run with the new adjusted intervals and increased max
retries in this commit allowed progress to continue past the 4th
checkpoint, but this time no retries were required at that point.
However, the browser tab crashed before reaching the 5th checkpoint.
The only other factor that changed was running with Chrome instead of
FireFox. Chrome is known to use more memory than FFox, so that could
be a limiting factor that cannot be avoided when using chrome. Also,
in my environment I have a FireFox profile config that specifies a
higher memory limit.