pause before counting all song elements
In the last two recent tests that used the new checkpoint test for page
load stalling, the song count at the failed checkpoint was equal to the
previous successful checkpoint. In other words, checking the total num
of songs loaded sometimes freezes the page. I'm not sure if the driver
is freezing or the web server of soundcloud is freezing. The driver
still accepts selenium commands, so it must be the web page, but I
don't know why that would cause the server to freeze. This checkpoint
is meant to heuristically determine whether the SC webserver has frozen
but we are only interacting ith the client webpage (via the browser).

So my first attempt to fix this is to simply pause for a relatively
long time before counting the songs at a checkpoint. I already was wary
of counting the songs too frequently because of the memory cost it can
incur, especially at higher levels, hence the checkpoint system rather
than every scroll cycle.