Increase pause between scrolls every checkpoint
The pause time should be increased even after a successful
checkpoint. The page load times always increase as more songs are loaded
so it is better to increase ahead of time. At a rate of +0.5, the
base_pause will increase by 1.5 by the 4th checkpoint, which is when
a stalled page can likely occur.

Also increased the pause after a failed checkpoint a little more.