Change scroll key to PAGE_DOWN from END
Sending PAGE_DOWN rather than END to the browser will trigger the lazy
loading list less often, and help prevent lag. When the lazy list is
fetching more songs it can take a long time when the list has grown to
be very long, and sending more actions to the browser from Selenium may
overload the driver and crash the browser. I need to slow down to let
the driver or browser, or server catch up to my requests.