Dynamically create checkpoints, increment basePause
I'm still adjusting these values as I test, but I noticed the webdriver
was increasingly getting slower as I hit the END key to scroll, forcing
a fetch of songs from the server. Using the PAGE_DOWN key will decrease
the frequency of fetches but because I think the webdriver is responding
sluggishly, not the SoundCloud server, I needed to increase the pause
between sending the PAGE_DOWN actionChain. It was getting sluggish after
only 70 or 80 END key scroll downs. Now it will reach 3500 PAGE_DOWNs
but the browser tab crashes randomly when it gets to that limit or
earlier. Or sometimes Selenium throws a TimeoutError when I'm using
Webdriver wait. In that case I can catch the error and save the reposts
collected so far. When the tab crashes I cannot.

Also noticed that Firefox webdriver is very sluggish after only 300
scrolls, taking over 1 min to respond and scroll again, even though the
incremented basePause should only add about 8 or 9 seconds by then.