Separate exception handling from grab first art
When an uncaught exception occurs in scrollReposts, it should be handled
separately. I also wanted to avoid calling scrollReposts twice, although
the intention was to call scrollReposts if I caught a TimeoutException
while trying to grab the first album art, which means I didn't get to
call it after successfully getting the first art.

The uncaught exception I'm getting is
'WebDriverException: Message: tab crashed'
Next, I will try to save the songs collected at the checkpoint in the
ul songs_list array in case the browser crashed and I get an Exception
when trying to save page source.