Save songs from checkpoint if webdriver fails
If an error occur while scrolling or the webdriver fails because the
tab crashed, we cannot save the page source from the driver. Hoping
to catch the exception when that happens and parse the songs saved at
the last checkpoint. Needed to refactor the transform function to
extract the ul unordered list of reposted songs directly because that
is the element captured at the checkpoint. It should still work when
saving the whole page source, because the element locator for should
be unique for BeautifulSoup to find it in either case.