save custom file name html, correctly save backup
The reason PurePath was imported was to easily save html using the url
to get the file name. No longer need PurePath now that we pass the file
name in as a parameter.

Did not realize save() was being called incorrectly for html backups in
error handling cases. Corrected those calls to save the page_source html
and use the new file name parameter.