Change scope of checkpoint_songs_list to local
An Selenium error occurred when trying to access the WebElement in the
finally block after scrollReposts finishes because the driver was closed
and quit. Selenium appears not to hold properties of WebElements in
memory and makes a URL request every time to get its text.

To save the html of the checkpoint_songs_list we must save it at
checkpoint intervals while scrolling.