Fix: save checkpoint html in BS4-parseable form
BS4=BeautifulSoup4 - The save() function wasn't being used correctly and
recently the function was refactored, so when calling it to save the
checkpoint_songs_list the arguments had to be rearranged.

The text property of a WebElement just gives the text inside the HTML
tags, not the html itself. The outerHTML attribute contains the songs
list wrapped in the ul element, which the transform() function expects.

The outerHTML needs to be wrapped in <html> tags so that BeautifulSoup
can parse it.

Now the checkpoint songs are successfully saved while scrolling, and if
the script is run with the checkpoint html as the first argument, it
correctly parses it to reposts-1.txt, skipping the scrolling and
scraping (as expected, that part was older and designed that way).