Get newestOldSong from previously scraped reposts
This will be used later to stop the script as soon as this song is
found, preventing the scraper from going any further than it needs
to in order to get only the newest songs since last run. The
drawback with this method is the entire collection of songs will not
be refreshed, thus dead links will not be pruned (dead links do not
exist on SC when scraping, so they will not exist in the new saved
list of songs).